requirements.txt
torch
numpy
requests
pycryptodome
